Moved helper functions to more appropriate controllers
