+++ gridt/controllers/helpers.py
@@ -1,7 +1,6 @@
 from contextlib import contextmanager
-from sqlalchemy import not_, desc
 from gridt.db import Session
-from gridt.models import User, MovementUserAssociation, Movement, Signal, Subscription
+from gridt.models import User, Movement
 from gridt import exc as GridtExceptions
 
 
@@ -22,71 +21,6 @@ def session_scope():
         raise
     finally:
         session.close()
-
-
-def leaders(user: User, movement: Movement, session: Session) -> list:
-    """
-    Create a query for the leaders of a user in a movement from a session.
-
-    :param gridt.models.user.User user: User that needs new leaders.
-    :param list exclude: List of users (can be a user model or an id) to
-    exclude from search.
-    :returns: List object
-    """
-    return [ mua.leader for mua in
-        session.query(MovementUserAssociation)
-        .filter(
-            MovementUserAssociation.follower_id == user.id,
-            MovementUserAssociation.movement_id == movement.id,
-            not_(MovementUserAssociation.leader_id.is_(None)),
-            MovementUserAssociation.destroyed.is_(None)
-        )
-    ]
-
-
-def _find_last_signal(
-    leader_id: int, movement_id: int, session: Session
-) -> Signal:
-    """Find the last signal the leader has sent to the movement."""
-    return (
-        session.query(Signal)
-        .filter_by(leader_id=leader_id, movement_id=movement_id)
-        .order_by(desc("time_stamp"))
-        .first()
-    )
-
-
-def extend_movement_json(movement, user, session):
-    movement_json = movement.to_json()
-    movement_json["subscribed"] = False
-
-    is_subscribed = (
-        session.query(Subscription)
-        .filter(
-            Subscription.user_id == user.id,
-            Subscription.movement_id == movement.id,
-            Subscription.time_removed.is_(None)
-        ).count()
-    )
-    
-    if is_subscribed:
-        movement_json["subscribed"] = True
-
-        last_signal = _find_last_signal(user.id, movement.id, session)
-        movement_json["last_signal_sent"] = (
-            last_signal.to_json() if last_signal else None
-        )
-
-        movement_json["leaders"] = []
-        for leader in leaders(user, movement, session):
-            leader_json = leader.to_json()
-
-            last_leader_signal = _find_last_signal(leader.id, movement.id, session)
-            if last_leader_signal:
-                leader_json.update(last_signal=last_leader_signal.to_json())
-
-            movement_json["leaders"].append(leader_json)
-    return movement_json
 
 
 def load_user(user_id: int, session: Session) -> User:
