+++ gridt/tests/controllers/test_subscription.py
@@ -50,17 +50,17 @@ class SubscriptionControllerUnitTest(BaseTest):
         user_id = user.id
         movement_id = movement.id
 
-        self.assertFalse(is_subscribed(user_id, movement_id))
+        self.assertFalse(is_subscribed(user_id, movement_id, self.session))
 
         subscription = Subscription(user, movement)
         self.session.add(subscription)
 
-        self.assertTrue(is_subscribed(user_id, movement_id))
+        self.assertTrue(is_subscribed(user_id, movement_id, self.session))
 
         subscription.unsubscribe()
         self.session.add(subscription)
 
-        self.assertFalse(is_subscribed(user_id, movement_id))
+        self.assertFalse(is_subscribed(user_id, movement_id, self.session))
 
     def test_new_subscription(self):
         user = self.create_user()
@@ -188,13 +188,13 @@ class SubscriptionControllerIntergrationTests(BaseTest):
         self.assertDictEqual(json_2['user'], andrei_json)
         self.assertEqual(json_2['time_started'], datetime(2023, 1, 6, 9, 30))
 
-        self.assertTrue(is_subscribed(antonin_id, movement_1_id))
-        self.assertTrue(is_subscribed(andrei_id, movement_1_id))
+        self.assertTrue(is_subscribed(antonin_id, movement_1_id, self.session))
+        self.assertTrue(is_subscribed(andrei_id, movement_1_id, self.session))
         self.assertIn(antonin_json, get_subscribers(movement_1_id))
         self.assertIn(andrei_json, get_subscribers(movement_1_id))
         self.assertEqual(get_subscriptions(antonin_id)[0]['id'], movement_1_id)
         self.assertEqual(get_subscriptions(andrei_id)[0]['id'], movement_1_id)
-        self.assertFalse(is_subscribed(antonin_id, movement_2_id))
+        self.assertFalse(is_subscribed(antonin_id, movement_2_id, self.session))
         self.assertNotIn(antonin_json, get_subscribers(movement_2_id))
         self.assertNotIn(andrei_json, get_subscribers(movement_2_id))
 
@@ -230,7 +230,7 @@ class SubscriptionControllerIntergrationTests(BaseTest):
         self.assertIsNone(json1['time_ended'])
         self.assertTrue(json1['subscribed'])
 
-        self.assertTrue(is_subscribed(antonin_id, movement_id))
+        self.assertTrue(is_subscribed(antonin_id, movement_id, self.session))
 
         with freeze_time("2023-01-06 10:00:00"):
             json2 = remove_subscription(antonin_id, movement_id)
@@ -240,7 +240,7 @@ class SubscriptionControllerIntergrationTests(BaseTest):
         self.assertEqual(json2['time_ended'], datetime(2023, 1, 6, 10, 0))
         self.assertFalse(json2['subscribed'])
 
-        self.assertFalse(is_subscribed(antonin_id, movement_id))
+        self.assertFalse(is_subscribed(antonin_id, movement_id, self.session))
         self.assertNotIn(antonin_json, get_subscribers(movement_id))
         self.assertListEqual([], get_subscriptions(antonin_id))
